chore: renamed agent_qa and agent_fkey in nexigraph
