--- ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_fkey/src/agent_fkey/__init__.py
@@ -1,2 +1,0 @@
-def main() -> None:
-    print("Hello from agent-fkey!")
