+++ ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_graph_gen/src/agent_graph_gen/__init__.py
@@ -1,0 +1,2 @@
+def main() -> None:
+    print("Hello from agent-graph-gen!")
