+++ ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_graph_gen/src/agent_graph_gen/tests/test_heuristics_e2e.py
@@ -2,7 +2,7 @@ import pytest
 import pytest_asyncio
 import logging
 import json
-from agent_fkey.relation_manager import RelationCandidateManager
+from agent_graph_gen.relation_manager import RelationCandidateManager
 
 from core.graph_db.neo4j.graph_db import Neo4jDB
 from core.utils import get_default_fresh_until, ObjEncoder
@@ -12,7 +12,7 @@ from core.models import RelationCandidate
 from core.models import Entity
 
 
-from agent_fkey.heuristics import HeuristicsProcessor
+from agent_graph_gen.heuristics import HeuristicsProcessor
 from core.constants import DEFAULT_LABEL, UPDATED_BY_KEY
