+++ ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_graph_qa/langgraph.json
@@ -1,7 +1,7 @@
 {
   "dependencies": ["."],
   "graphs": {
-    "qna_agent": "./src/agent_qa/agent.py:graph"
+    "qna_agent": "./src/agent_graph_qa/agent.py:graph"
   },
   "env": ".env"
 }
