+++ ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_graph_qa/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "agent-fkey"
+name = "agent-graph-qa"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
@@ -8,17 +8,22 @@ authors = [
 ]
 requires-python = ">=3.13"
 dependencies = [
+    "a2a-sdk>=0.2.8",
+    "click>=8.2.1",
     "core",
-    "langchain>=0.3.26",
+    "httpx>=0.28.1",
+    "langchain>=0.3.25",
+    "langchain-aws>=0.2.26",
     "langchain-openai>=0.3.23",
     "langgraph>=0.4.8",
-    "pytest>=8.4.1",
-    "pytest-asyncio>=1.1.0",
+    "langmem>=0.0.27",
+    "pydantic>=2.11.7",
     "python-dotenv>=1.1.0",
+    "uvicorn>=0.34.3",
 ]
 
 [project.scripts]
-agent-fkey = "agent_fkey:main"
+agent-graph-qa = "agent_graph_qa:main"
 
 [build-system]
 requires = ["hatchling"]
