+++ ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_graph_qa/src/agent_graph_qa/a2a_executor.py
@@ -17,7 +17,7 @@ from a2a.utils import (
 )
 from a2a.utils.errors import ServerError
 
-from agent_qa.agent import QnAAgent
+from agent_graph_qa.agent import QnAAgent
 from core import utils
 
 logger = utils.get_logger(__name__)
