--- ai_platform_engineering/knowledge_bases/graph_rag/nexigraph/nexigraph/agent_qa/src/agent_qa/__init__.py
@@ -1,2 +1,0 @@
-def main() -> None:
-    print("Hello from agent-qa!")
